deps: make react peer dep optional and remove react import from core types
Move FormoAnalyticsProviderProps (and its ReactNode dependency) out of
the framework-agnostic types/base.ts and into FormoAnalyticsProvider.tsx,
and mark react / @types/react as optional peerDependencies. This lets
non-React consumers (e.g. Angular apps using formofy() or
FormoAnalytics.init() directly) install and type-check the package
without pulling react.
