+++ src/types/base.ts
@@ -6,7 +6,6 @@ import {
   TransactionStatus,
 } from "./events";
 import { EIP1193Provider } from "./provider";
-import { ReactNode } from "react";
 import { SolanaOptions } from "../solana/types";
 
 export type Nullable<T> = T | null;
@@ -285,9 +284,3 @@ export interface Options {
   ready?: (formo: IFormoAnalytics) => void;
 }
 
-export interface FormoAnalyticsProviderProps {
-  writeKey: string;
-  options?: Options;
-  disabled?: boolean;
-  children: ReactNode;
-}
